requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.8.0
spotipy==2.23.0
python-dotenv==1.0.0
